wip(#1101): extract lower/decorators
